requirements.txt
python-telegram-bot==21.5
supabase==2.9.1
google-generativeai==0.8.3
groq==0.11.0
APScheduler==3.10.4
httpx==0.27.2
